apps/api/pyproject.toml
[project]
name = "tocky-api"
version = "0.1.0"
description = "Tocky API service"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "pydantic-settings",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "httpx",
    "oss2",
    "pyjwt",
]

[dependency-groups]
dev = [
    "ruff",
    "ty",
    "pytest",
    "httpx",
]

[tool.fastapi]
entrypoint = "app.main:app"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "FAST", # FastAPI-specific rules
]
